+++ Controller/Router.php
@@ -3,12 +3,11 @@
 namespace Conekta\Payments\Controller;
 
 use Magento\Framework\App\ActionFactory;
-use Magento\Framework\App\ActionInterface;
 use Magento\Framework\App\RequestInterface;
 use Magento\Framework\App\ResponseInterface;
 use Magento\Framework\App\RouterInterface;
 use Conekta\Payments\Helper\Data as ConektaHelper;
-use Conekta\Payments\Logger\Logger as ConektaLogger;
+use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\Url;
 
 class Router implements RouterInterface
@@ -16,46 +15,39 @@ class Router implements RouterInterface
     /**
      * @var ActionFactory
      */
-    protected $actionFactory;
+    protected ActionFactory $actionFactory;
 
     /**
      * @var ResponseInterface
      */
-    protected $_response;
+    protected ResponseInterface $_response;
 
     /**
      * @var ConektaHelper
      */
-    private $_conektaHelper;
+    private ConektaHelper $_conektaHelper;
 
-    /**
-     * @var ConektaLogger
-     */
-    private $_conektaLogger;
 
     /**
      * @param ActionFactory $actionFactory
      * @param ResponseInterface $response
      * @param ConektaHelper $conektaHelper
-     * @param ConektaLogger $conektaLogger
      */
     public function __construct(
         ActionFactory $actionFactory,
         ResponseInterface $response,
-        ConektaHelper $conektaHelper,
-        ConektaLogger $conektaLogger
+        ConektaHelper $conektaHelper
     ) {
         $this->actionFactory = $actionFactory;
         $this->_response = $response;
         $this->_conektaHelper = $conektaHelper;
-        $this->_conektaLogger = $conektaLogger;
     }
 
     /**
      * Validate and Match
      *
      * @param RequestInterface $request
-     * @return ActionInterface
+     * @throws NoSuchEntityException
      */
     public function match(RequestInterface $request)
     {
@@ -73,8 +65,6 @@ class Router implements RouterInterface
         if ($pathRequest === $pathWebhook) {
             $request->setModuleName('conekta')->setControllerName('webhook')->setActionName('index');
             $request->setAlias(Url::REWRITE_REQUEST_PATH_ALIAS, $pathRequest);
-        } else {
-            return;
         }
     }
 }
